test(metal): add msm wrapper test on metal implementation
