+++ mopro-core/src/middleware/gpu_explorations/metal/tests/mod.rs
@@ -1,2 +1,2 @@
 pub mod test_bn254;
-pub mod test_msm_accumulation;
+pub mod test_msm;
